refactor: rename file with user types
